requirements.txt
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
bandit==1.9.4
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.6
cryptography==46.0.6
cyclonedx-python-lib==11.7.0
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
filelock==3.25.2
google-auth==2.49.1
google-genai==1.68.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.18
idna==3.11
iniconfig==2.3.0
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
nodeenv==1.10.0
packageurl-python==0.17.6
packaging==26.0
pillow==12.1.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.4
pluggy==1.6.0
py-serializable==2.1.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
pytesseract==0.3.13
pytest==9.0.2
python-discovery==1.2.1
python-dotenv==1.2.2
python-telegram-bot==22.7
PyYAML==6.0.3
requests==2.33.0
rich==14.3.3
ruff==0.15.8
sniffio==1.3.1
sortedcontainers==2.4.0
stevedore==5.7.0
tenacity==9.1.4
tomli==2.4.1
tomli_w==1.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
virtualenv==21.2.0
websockets==16.0
